MicrosoftFantasyBroadcaster/ScraperService/requirements.txt
espn_api>=0.3.1
pika>=1.3.1
python-dotenv>=1.0.0
openai>=1.12.0
requests>=2.31.0